Add tests for new WebhookPlugin methods
