+++ saleor/plugins/webhook/tests/conftest.py
@@ -1,5 +1,10 @@
 import pytest
 
+from saleor.app.models import App
+from saleor.plugins.manager import get_plugins_manager
+from saleor.webhook.event_types import WebhookEventType
+from saleor.webhook.models import Webhook
+
 
 @pytest.fixture
 def tax_line_data_response():
@@ -25,3 +30,43 @@ def tax_data_response(tax_line_data_response):
         "shipping_price_net_amount": 12.34,
         "lines": [tax_line_data_response] * 5,
     }
+
+
+@pytest.fixture
+def tax_app(permission_handle_taxes):
+    app = App.objects.create(name="Sample app objects", is_active=True)
+    app.tokens.create(name="Default")
+    app.permissions.add(permission_handle_taxes)
+    return app
+
+
+@pytest.fixture
+def tax_checkout_webhook(tax_app):
+    webhook = Webhook.objects.create(
+        name="Tax checkout webhook",
+        app=tax_app,
+        target_url="https://www.example.com/tax-checkout",
+    )
+    webhook.events.create(event_type=WebhookEventType.CHECKOUT_CALCULATE_TAXES)
+    return webhook
+
+
+@pytest.fixture
+def tax_order_webhook(tax_app):
+    webhook = Webhook.objects.create(
+        name="Tax order webhook",
+        app=tax_app,
+        target_url="https://www.example.com/tax-order",
+    )
+    webhook.events.create(event_type=WebhookEventType.ORDER_CALCULATE_TAXES)
+    return webhook
+
+
+@pytest.fixture
+def webhook_plugin(settings):
+    def factory():
+        settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
+        manager = get_plugins_manager()
+        return manager.global_plugins[0]
+
+    return factory
