+++ saleor/plugins/webhook/tests/test_tax_webhook.py
@@ -1,65 +1,18 @@
+from unittest import mock
 from unittest.mock import Mock
 
-import pytest
-from django.contrib.auth.models import Permission
-
-from saleor.core.permissions import CheckoutPermissions
-from saleor.plugins.webhook.tasks import trigger_tax_webhook_sync
 from saleor.plugins.webhook.utils import parse_tax_data
 from saleor.webhook.event_types import WebhookEventType
-from saleor.webhook.models import Webhook
+from saleor.webhook.payloads import generate_checkout_payload, generate_order_payload
 
 
-@pytest.fixture
-def tax_app(app):
-    app.permissions.add(
-        Permission.objects.filter(
-            codename=CheckoutPermissions.HANDLE_TAXES.codename
-        ).first()
-    )
-    return app
-
-
-@pytest.fixture
-def tax_checkout_webhook(tax_app):
-    webhook = Webhook.objects.create(
-        name="Tax checkout webhook",
-        app=tax_app,
-        target_url="https://www.example.com/tax-checkout",
-    )
-    webhook.events.create(event_type=WebhookEventType.CHECKOUT_CALCULATE_TAXES)
-    return webhook
-
-
-@pytest.fixture
-def tax_checkout_webhooks(tax_app):
-    webhooks = []
-    for i in range(8):
-        webhook = Webhook.objects.create(
-            name="Tax checkout webhook",
-            app=tax_app,
-            target_url="https://www.example.com/tax-checkout",
-        )
-        webhook.events.create(event_type=WebhookEventType.CHECKOUT_CALCULATE_TAXES)
-        webhooks.append(webhook)
-
-    return webhooks
-
-
-@pytest.fixture
-def tax_order_webhook(tax_app):
-    webhook = Webhook.objects.create(
-        name="Tax order webhook",
-        app=tax_app,
-        target_url="https://www.example.com/tax-order",
-    )
-    webhook.events.create(event_type=WebhookEventType.ORDER_CALCULATE_TAXES)
-    return webhook
-
-
-def test_trigger_tax_webhook_sync(
+def test_get_taxes_for_checkout(
+    permission_handle_taxes,
+    webhook_plugin,
     tax_checkout_webhook,
     tax_data_response,
+    checkout,
+    tax_app,
     monkeypatch,
 ):
     # given
@@ -67,73 +20,83 @@ def test_trigger_tax_webhook_sync(
     monkeypatch.setattr(
         "saleor.plugins.webhook.tasks.send_webhook_request_sync", mock_request
     )
-    event_type = WebhookEventType.CHECKOUT_CALCULATE_TAXES
-    data = '{"key": "value"}'
+    plugin = webhook_plugin()
 
     # when
-    tax_data = trigger_tax_webhook_sync(event_type, data)
+    tax_data = plugin.get_taxes_for_checkout(checkout, None)
 
     # then
     mock_request.assert_called_once_with(
         tax_checkout_webhook.pk,
         tax_checkout_webhook.target_url,
         tax_checkout_webhook.secret_key,
-        event_type,
-        data,
+        WebhookEventType.CHECKOUT_CALCULATE_TAXES,
+        generate_checkout_payload(checkout),
     )
     assert tax_data == parse_tax_data(tax_data_response)
 
 
-@pytest.mark.parametrize("no_of_unsuccessful_webhooks", [0, 4, 7])
-def test_trigger_tax_webhook_sync_multiple_webhooks(
-    tax_checkout_webhooks,
-    tax_data_response,
+@mock.patch("saleor.plugins.webhook.tasks.send_webhook_request_sync")
+def test_get_taxes_for_checkout_no_permission(
+    mock_request,
+    webhook_plugin,
+    checkout,
     monkeypatch,
-    no_of_unsuccessful_webhooks,
 ):
     # given
-    def new_send_webhook_request_sync(pk, *args):
-        if pk == tax_checkout_webhooks[no_of_unsuccessful_webhooks].pk:
-            return tax_data_response
-        else:
-            return {}
+    plugin = webhook_plugin()
 
-    mock_request = Mock(wraps=new_send_webhook_request_sync)
+    # when
+    tax_data = plugin.get_taxes_for_checkout(checkout, None)
+
+    # then
+    assert mock_request.call_count == 0
+    assert tax_data is None
+
+
+def test_get_taxes_for_order(
+    permission_handle_taxes,
+    webhook_plugin,
+    tax_order_webhook,
+    tax_data_response,
+    order,
+    tax_app,
+    monkeypatch,
+):
+    # given
+    mock_request = Mock(return_value=tax_data_response)
     monkeypatch.setattr(
         "saleor.plugins.webhook.tasks.send_webhook_request_sync", mock_request
     )
-    event_type = WebhookEventType.CHECKOUT_CALCULATE_TAXES
-    data = '{"key": "value"}'
+    plugin = webhook_plugin()
 
     # when
-    tax_data = trigger_tax_webhook_sync(event_type, data)
+    tax_data = plugin.get_taxes_for_order(order, None)
 
     # then
-    assert mock_request.call_count == no_of_unsuccessful_webhooks + 1
-    for call, webhook in zip(mock_request.call_args_list, tax_checkout_webhooks):
-        assert call == (
-            (webhook.pk, webhook.target_url, webhook.secret_key, event_type, data),
-            {},
-        )
+    mock_request.assert_called_once_with(
+        tax_order_webhook.pk,
+        tax_order_webhook.target_url,
+        tax_order_webhook.secret_key,
+        WebhookEventType.ORDER_CALCULATE_TAXES,
+        generate_order_payload(order),
+    )
     assert tax_data == parse_tax_data(tax_data_response)
 
 
-def test_trigger_tax_webhook_sync_invalid_webhooks(
-    tax_checkout_webhooks,
-    tax_data_response,
+@mock.patch("saleor.plugins.webhook.tasks.send_webhook_request_sync")
+def test_get_taxes_for_order_no_permission(
+    mock_request,
+    webhook_plugin,
+    order,
     monkeypatch,
 ):
     # given
-    mock_request = Mock(return_value={})
-    monkeypatch.setattr(
-        "saleor.plugins.webhook.tasks.send_webhook_request_sync", mock_request
-    )
-    event_type = WebhookEventType.CHECKOUT_CALCULATE_TAXES
-    data = '{"key": "value"}'
+    plugin = webhook_plugin()
 
     # when
-    tax_data = trigger_tax_webhook_sync(event_type, data)
+    tax_data = plugin.get_taxes_for_order(order, None)
 
     # then
-    assert mock_request.call_count == len(tax_checkout_webhooks)
+    assert mock_request.call_count == 0
     assert tax_data is None
